Added routine for isolating Augmenter-genned sites.
Small reorganization of underlying map content. Removed Map
abstract class because it is no longer of use with TMaps,
added _Taggable class to a basic subset of its behavior. LinearMap
objects no longer accept unused arguments for compatibility.

This should not change user experiences, but will break tests
that used the Map interface for generality. Tests will have to be
fixed in a different commit.